+++ dflib-parquet/src/main/java/org/dflib/parquet/read/SchemaProjector.java
@@ -9,8 +9,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-
-
 @FunctionalInterface
 public interface SchemaProjector {